pyproject.toml
[tool.poetry]
name = "icon-caster"
version = "0.1.0"
description = "A CLI image to .ico converter made for my personal use."
authors = ["mooncell07 <mooncell07@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
